+++ addon/components/t-scroll.js
@@ -1,9 +1,5 @@
-import Ember from 'ember';
+import Component from '@ember/component';
 import layout from '../templates/components/t-scroll';
-
-const {
-  Component
-} = Ember;
 
 /**
   Transformicon Scroll Indicator component.