+++ tests/dummy/app/app.js
@@ -1,12 +1,8 @@
-import Ember from 'ember';
+import LinkComponent from '@ember/routing/link-component';
+import Application from '@ember/application';
 import Resolver from './resolver';
 import loadInitializers from 'ember-load-initializers';
 import config from './config/environment';
-
-const {
-  LinkComponent,
-  Application
-} = Ember;
 
 LinkComponent.reopen({
   attributeBindings: ['role']