+++ tests/helpers/module-for-acceptance.js
@@ -1,10 +1,8 @@
+import { resolve } from 'rsvp';
 import { module } from 'qunit';
-import Ember from 'ember';
 import startApp from '../helpers/start-app';
 import destroyApp from '../helpers/destroy-app';
 import 'dummy/tests/helpers/percy/register-helpers';
-
-const { RSVP: { resolve } } = Ember;
 
 /* eslint-disable ember/named-functions-in-promises, consistent-return */
 export default function(name, options = {}) {